Remove rng. Cleans up workflow.
