+++ parachain/pallets/ethereum-beacon-client/fuzz/Cargo.toml
@@ -14,7 +14,6 @@ snowbridge-ethereum-beacon-client = {path = "..", features = [ "fuzzing" ]}
 snowbridge-beacon-primitives= {path = "../../../primitives/beacon"}
 hex-literal = { version = "0.4.1" }
 sp-core = { git = "https://github.com/paritytech/substrate.git", branch = "master", default-features = false }
-rand = "0.8"
 
 # Prevent this from interfering with workspaces
 [workspace]
